Refatorando rotas, paginas de player e home e estrutura do projeto
