+++ src/App.js
@@ -1,20 +1,30 @@
 import React, { Component } from 'react';
 import { BrowserRouter, Switch, Route } from 'react-router-dom';
-import Home from './pages/Home';
-import VideoInfos from './pages/VideoInfos';
-import Player from './pages/Player';
-import Profile from './pages/Profile';
+
+import * as ROUTES from './constants/routes';
+
+import NavBar from './components/NavBar/NavBar';
+import Footer from './components/Footer/Footer';
+
+import Home from './pages/Home/Home';
+import VideoInfos from './pages/VideoInfo/VideoInfos';
+import Player from './pages/Player/Player';
+import Profile from './pages/Profile/Profile';
 
 class App extends Component {
   render() {
     return (
       <BrowserRouter>
+        <NavBar />
+
         <Switch>
-          <Route path="/" exact component={Home} />
-          <Route path="/Video" exact component={VideoInfos} />
-          <Route path="/Player" exact component={Player} />
-          <Route path="/Profile" exact component={Profile} />
+          <Route path={ROUTES.HOME} exact component={Home} />
+          <Route path={ROUTES.VIDEO} exact component={VideoInfos} />
+          <Route path={ROUTES.PLAYER} exact component={Player} />
+          <Route path={ROUTES.PROFILE} exact component={Profile} />
         </Switch>
+
+        <Footer />
       </BrowserRouter>
     );
   }
