+++ src/components/VideoPlayer/VideoPlayer.css
@@ -1,0 +1,37 @@
+iframe {
+  width: 97vw;
+  height: 40vh;
+  margin-left: -21px;
+  margin-top: -20px;
+  position: fixed;
+}
+
+.title {
+  margin-top: 0;
+  padding-top: 40vh;
+}
+
+@media only screen and (min-width: 1224px) {
+  iframe {
+    position: relative;
+    margin-left: 0;
+    margin-top: 0;
+    width: 100%;
+    height: 60vh;
+    grid-area: video;
+  }
+  .Views-Film {
+    grid-area: views;
+  }
+  .title {
+    padding-top: 10px;
+    grid-area: title;
+  }
+  .videoPlayer {
+    display: grid;
+    grid-template-columns: 2fr 210px;
+    grid-template-rows: 60vh 20px 50px;
+    grid-template-areas: 'video video' 'title actions' 'views actions';
+    grid-gap: 15px;
+  }
+}
