+++ src/constants/routes.js
@@ -1,0 +1,4 @@
+export const HOME = '/home';
+export const VIDEO = '/video';
+export const PLAYER = `${VIDEO}/player`;
+export const PROFILE = '/profile';
