+++ src/pages/Player/Player.css
@@ -1,0 +1,31 @@
+.containerLeft {
+  margin: 0;
+  grid-area: containerLeft;
+}
+
+.containerRight {
+  grid-area: containerRight;
+}
+
+.containerBottom {
+  grid-area: containerBottom;
+}
+
+@keyframes textComment {
+  0% {
+    height: 15px;
+  }
+  100% {
+    height: 60px;
+  }
+}
+
+@media only screen and (min-width: 1224px) {
+  .containerPlayer {
+    display: grid;
+    grid-template-areas: 'containerLeft containerRight' 'containerBottom containerRight';
+    grid-template-columns: 1fr 380px;
+    grid-template-rows: 65vh 1fr;
+    grid-gap: 80px;
+  }
+}
