--- src/pages/styles/Player.css
@@ -1,81 +1,0 @@
-.containerLeft{
-    margin: 0;
-    grid-area: containerLeft;
-}
-
-.containerRight{
-    grid-area: containerRight;
-}
-
-.containerBottom{
-    grid-area: containerBottom;
-}
-
-.containerPlayer .actionDescription{
-    display: none;
-}
-
-.containerPlayer .ActionsPlayer{
-    justify-content: left;
-}
-
-iframe{
-    width: 97vw;
-    height: 40vh;
-    margin-left: -21px;
-    margin-top: -20px;
-    position: fixed;
-}
-
-.title{
-    margin-top: 0;
-    padding-top: 40vh;
-}
-
-@keyframes textComment {
-    0%   {height: 15px;}
-    100% {height: 60px;}
-}
-
-@media only screen and (min-width : 1224px) {
-    iframe{
-        position: relative;
-        margin-left: 0;
-        margin-top: 0;
-        width: 100%;
-        height: 60vh;
-        grid-area: video;
-    }
-    .Views-Film{
-        grid-area: views;
-    }
-    .title{
-        padding-top: 10px;
-        grid-area: title;
-    }
-    .actionsPlayer{
-        grid-area: actions;
-    }
-    .containerPlayer .ActionsPlayer{
-        margin-top: 25px;
-        flex-direction: row-reverse;
-    }
-    .containerPlayer .ActionsPlayer .applauses{
-        flex-direction: row-reverse;
-        text-align: right;
-    }
-    .containerPlayer{
-        display: grid;
-        grid-template-areas: 'containerLeft containerRight' 'containerBottom containerRight';
-        grid-template-columns: 1fr 380px;
-        grid-template-rows: 65vh 1fr;
-        grid-gap: 80px;
-    }
-    .containerLeft{
-        display: grid;
-        grid-template-columns: 2fr 210px;
-        grid-template-rows: 60vh 20px 50px;
-        grid-template-areas: "video video" "title actions" "views actions";
-        grid-gap: 15px;
-    }
-}
